requirements.txt
amqp==5.1.1
asgiref==3.6.0
async-timeout==4.0.2
billiard==3.6.4.0
celery==5.2.7
cffi==1.15.1
channels==4.0.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cryptography==40.0.2
Django==4.2
django-celery-results==2.5.0
django-debug-toolbar==4.0.0
django-phonenumber-field==7.0.2
django-redis==5.2.0
django-rest-knox==4.2.0
djangorestframework==3.14.0
jsonfield==3.1.0
kombu==5.2.4
mqttasgi==1.2.1
paho-mqtt==1.6.1
phonenumberslite==8.13.10
prompt-toolkit==3.0.38
pycparser==2.21
python-dotenv==1.0.0
pytz==2023.3
redis==4.5.4
six==1.16.0
sqlparse==0.4.4
vine==5.0.0
wcwidth==0.2.6
